requirements.txt
absl-py==0.13.0
ai2thor==1.0.1
astunparse==1.6.3
blis==0.7.4
cachetools==4.2.2
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.2
click==8.0.1
cryptography==3.4.7
cycler==0.10.0
cymem==2.0.5
Flask==2.0.1
gast==0.3.3
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.38.1
h5py==2.10.0
idna==3.2
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib
<<<<<<< HEAD
ai2thor==2.0.0
Cython
scikit-image
h5py
tensorboardX
torchvision
keras
tqdm
gputil
opencv-python
scipy
networkx
gensim
spacy
tensorflow
torch==1.6.0
=======
murmurhash==1.0.5
numpy
oauthlib==3.1.1
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.2
Pillow==8.0.1
pip==21.1.3
plac==1.1.3
preshed==3.0.5
progressbar2==3.53.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-utils==2.5.6
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
scikit-image==0.17.2 
setuptools
six==1.16.0
spacy==2.3.5
srsly==1.0.5
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.2.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tensorflow-gpu==2.1.0
thinc==7.4.5
torch==1.7.1
torchvision==0.8.2
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
wheel==0.36.2
wrapt==1.12.1
zipp==3.4.0
>>>>>>> 656105dc0f394b0599d8906597b6bfeb3fb52d5c
